lambda_sentiment/requirements.txt
boto3==1.28.85
psycopg2-binary==2.9.9
